+++ spec/helpers/spy_input.h
@@ -12,14 +12,14 @@ struct SpyInputEdit {
 };
 
 class SpyInput {
-  size_t chars_per_chunk;
-  size_t buffer_size;
+  uint32_t chars_per_chunk;
+  uint32_t buffer_size;
   char *buffer;
-  size_t byte_offset;
+  uint32_t byte_offset;
   std::vector<SpyInputEdit> undo_stack;
 
-  static const char * read(void *, size_t *);
-  static int seek(void *, size_t, size_t);
+  static const char * read(void *, uint32_t *);
+  static int seek(void *, uint32_t, uint32_t);
   std::pair<std::string, TSPoint> swap_substr(size_t, size_t, std::string);
 
  public:
