+++ src/runtime/lexer.h
@@ -17,10 +17,10 @@ typedef struct {
   Length token_start_position;
 
   const char *chunk;
-  size_t chunk_start;
-  size_t chunk_size;
+  uint32_t chunk_start;
+  uint32_t chunk_size;
 
-  size_t lookahead_size;
+  uint32_t lookahead_size;
 
   TSInput input;
   TSLogger logger;
